+++ xhci.device/include/xhci/xhci.h
@@ -24,6 +24,7 @@
 #include <proto/exec.h>
 #endif
 
+#include <exec/execbase.h> /* DMA_ReadFromRAM for CachePreDMA(); older NDKs don't pull it in transitively */
 #include <bits.h>
 #include <iomem.h>
 #include <slab.h>
@@ -169,12 +170,12 @@ struct xhci_ctrl
 /* Pre-DMA flush for a buffer.  @flags is passed straight to CachePreDMA: 0 for a
  * plain clean+invalidate, or DMA_ReadFromRAM for an OUT buffer (device reads RAM)
  * which only needs a clean. */
-inline void xhci_flush_cache(void *addr, u32 len, ULONG flags)
+inline void xhci_flush_cache(void *addr, ULONG len, ULONG flags)
 {
 	CachePreDMA((APTR)addr, &len, flags);
 }
 
-inline void xhci_inval_cache(void *addr, u32 len)
+inline void xhci_inval_cache(void *addr, ULONG len)
 {
 	CachePostDMA((APTR)addr, &len, 0);
 }
